Added new Uniform alias types. Re-organized alias type package. Updated source files to import all aliases from prelude.
