+++ deep_causality/src/types/alias_types/alias_primitives.rs
@@ -1,0 +1,12 @@
+/*
+ * SPDX-License-Identifier: MIT
+ * Copyright (c) "2025" . The DeepCausality Authors and Contributors. All Rights Reserved.
+ */
+
+pub type IdentificationValue = u64;
+
+pub type DescriptionValue = String;
+pub type NumericalValue = f64;
+
+pub type NumberType = u64;
+pub type FloatType = f64;
