+++ deep_causality/src/types/alias_types/mod.rs
@@ -2,119 +2,8 @@
  * SPDX-License-Identifier: MIT
  * Copyright (c) "2025" . The DeepCausality Authors and Contributors. All Rights Reserved.
  */
-
-use crate::prelude::{
-    BaseSymbol, CausalityError, Causaloid, CausaloidGraph, Context, Contextoid, Data,
-    EuclideanSpace, EuclideanSpacetime, EuclideanTime, Model, SymbolicResult,
-};
-use std::collections::HashMap;
-use std::sync::{Arc, RwLock};
-
-// Type aliases
-pub type IdentificationValue = u64;
-pub type NumericalValue = f64;
-pub type DescriptionValue = String;
-
-// Thread safe Interior mutability in Rust
-// https://ricardomartins.cc/2016/06/25/interior-mutability-thread-safety
-pub type ArcRWLock<T> = Arc<RwLock<T>>;
-
-// Fn aliases for assumable, assumption, & assumption collection
-pub type EvalFn = fn(&[NumericalValue]) -> bool;
-
-pub type SymbolicRepr = String;
-pub type SymbolicCausalFn = fn(SymbolicRepr) -> Result<SymbolicResult, CausalityError>;
-
-pub type ProbabilisticCausalFn = fn(NumericalValue) -> Result<NumericalValue, CausalityError>; // Bayes update, etc.
-
-// Fn aliases for causal function with and without context
-pub type CausalFn = fn(&NumericalValue) -> Result<bool, CausalityError>;
-
-pub type ContextualCausalDataFn<D, S, T, ST, SYM, VS, VT> =
-    fn(&NumericalValue, &Arc<Context<D, S, T, ST, SYM, VS, VT>>) -> Result<bool, CausalityError>;
-
-pub type ContextualCausalFn<D, S, T, ST, SYM, VS, VT> =
-    fn(&Arc<Context<D, S, T, ST, SYM, VS, VT>>) -> Result<bool, CausalityError>;
-
-// Default type aliases for basic causaloids
-
-pub type BaseNumberType = u64;
-pub type BaseFloatType = f64;
-
-pub type BaseModel = Model<
-    Data<BaseNumberType>,
-    EuclideanSpace,
-    EuclideanTime,
-    EuclideanSpacetime,
-    BaseSymbol,
-    BaseFloatType,
-    BaseFloatType,
->;
-
-pub type BaseCausaloid = Causaloid<
-    Data<BaseNumberType>,
-    EuclideanSpace,
-    EuclideanTime,
-    EuclideanSpacetime,
-    BaseSymbol,
-    BaseFloatType,
-    BaseFloatType,
->;
-
-pub type BaseCausaloidVec = Vec<
-    Causaloid<
-        Data<BaseNumberType>,
-        EuclideanSpace,
-        EuclideanTime,
-        EuclideanSpacetime,
-        BaseSymbol,
-        BaseFloatType,
-        BaseFloatType,
-    >,
->;
-
-pub type BaseCausalMap = HashMap<
-    usize,
-    Causaloid<
-        Data<BaseNumberType>,
-        EuclideanSpace,
-        EuclideanTime,
-        EuclideanSpacetime,
-        BaseSymbol,
-        BaseFloatType,
-        BaseFloatType,
-    >,
->;
-
-pub type BaseCausalGraph = CausaloidGraph<
-    Causaloid<
-        Data<BaseNumberType>,
-        EuclideanSpace,
-        EuclideanTime,
-        EuclideanSpacetime,
-        BaseSymbol,
-        BaseFloatType,
-        BaseFloatType,
-    >,
->;
-
-// Default type alias for basic context. It's used in tests
-pub type BaseContext = Context<
-    Data<BaseNumberType>,
-    EuclideanSpace,
-    EuclideanTime,
-    EuclideanSpacetime,
-    BaseSymbol,
-    BaseFloatType,
-    BaseFloatType,
->;
-
-pub type BaseContextoid = Contextoid<
-    Data<BaseNumberType>,
-    EuclideanSpace,
-    EuclideanTime,
-    EuclideanSpacetime,
-    BaseSymbol,
-    BaseFloatType,
-    BaseFloatType,
->;
+pub mod alias_function;
+pub mod alias_lock;
+pub mod alias_primitives;
+pub mod aliases_base;
+pub mod aliases_uniform;
